Removed c2w package layer for sim control interactions.
Removed c2w package layer for sim control interactions.
