+++ src/java/main/gov/nist/ucef/hla/example/fedman/FedManFederate.java
@@ -38,9 +38,9 @@ import gov.nist.ucef.hla.example.util.Constants;
 import gov.nist.ucef.hla.example.util.StringUtils;
 import gov.nist.ucef.hla.example.util.cmdargs.ArgException;
 import gov.nist.ucef.hla.ucef.NoOpFederate;
-import gov.nist.ucef.hla.ucef.interaction.c2w.SimEnd;
-import gov.nist.ucef.hla.ucef.interaction.c2w.SimPause;
-import gov.nist.ucef.hla.ucef.interaction.c2w.SimResume;
+import gov.nist.ucef.hla.ucef.SimEnd;
+import gov.nist.ucef.hla.ucef.SimPause;
+import gov.nist.ucef.hla.ucef.SimResume;
 
 /**
  *		            ___
