+++ src/java/main/gov/nist/ucef/hla/example/ucef/UCEFPingFederate.java
@@ -34,11 +34,11 @@ import gov.nist.ucef.hla.base.UCEFException;
 import gov.nist.ucef.hla.base.UCEFSyncPoint;
 import gov.nist.ucef.hla.example.util.Constants;
 import gov.nist.ucef.hla.example.util.FileUtils;
+import gov.nist.ucef.hla.ucef.FederateJoin;
+import gov.nist.ucef.hla.ucef.SimEnd;
+import gov.nist.ucef.hla.ucef.SimPause;
+import gov.nist.ucef.hla.ucef.SimResume;
 import gov.nist.ucef.hla.ucef.UCEFFederateBase;
-import gov.nist.ucef.hla.ucef.interaction.c2w.FederateJoin;
-import gov.nist.ucef.hla.ucef.interaction.c2w.SimEnd;
-import gov.nist.ucef.hla.ucef.interaction.c2w.SimPause;
-import gov.nist.ucef.hla.ucef.interaction.c2w.SimResume;
 import hla.rti1516e.encoding.EncoderFactory;
 
 /**
